Set specific version for `postgresql-client-9.6` (#174)
A new version was introduced in the repository, making the `apt-get
install` call fail. Besides, removed empty files and added some checks
to the `setup.sh` scripts to make them install dependencies only if a
dependency definition file exists.

+++ aether-odk-module/conf/docker/setup.sh
@@ -14,12 +14,15 @@ export PYTHONUNBUFFERED=true
 # install packages
 ################################################################################
 
-# Add postgres apt repo to get more recent postgres versions
-echo 'deb http://apt.postgresql.org/pub/repos/apt/ jessie-pgdg main' > /etc/apt/sources.list.d/pgdg.list
-wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | apt-key add -
+PACKAGE_LIST=/tmp/apt-packages.txt
+if [ -f "$PACKAGE_LIST" ]; then
+    # Add postgres apt repo to get more recent postgres versions
+    echo 'deb http://apt.postgresql.org/pub/repos/apt/ jessie-pgdg main' > /etc/apt/sources.list.d/pgdg.list
+    wget --quiet -O - https://www.postgresql.org/media/keys/ACCC4CF8.asc | apt-key add -
 
-apt-get update -qq
-cat /tmp/apt-packages.txt | xargs apt-get -qq --yes --force-yes install
+    apt-get update -qq
+    apt-get -qq --yes --force-yes install `cat $PACKAGE_LIST`
+fi
 
 # upgrade pip
 pip install --upgrade pip
